+++ src/alloc.c
@@ -7,11 +7,8 @@
 
 #include "alloc.h"
 
-#if defined(GIT_MSVC_CRTDBG)
-# include "win32/w32_crtdbg_stacktrace.h"
-#else
-# include "allocators/stdalloc.h"
-#endif
+#include "allocators/stdalloc.h"
+#include "allocators/win32_crtdbg.h"
 
 git_allocator git__allocator;
 
@@ -44,12 +41,3 @@ int git_allocator_setup(git_allocator *allocator)
 	memcpy(&git__allocator, allocator, sizeof(*allocator));
 	return 0;
 }
-
-#if !defined(GIT_MSVC_CRTDBG)
-int git_win32_crtdbg_init_allocator(git_allocator *allocator)
-{
-	GIT_UNUSED(allocator);
-	git_error_set(GIT_EINVALID, "crtdbg memory allocator not available");
-	return -1;
-}
-#endif
